Solve day 2
I did part 1 in Uiua first (see next commit) which was not made to
easily solve part 2. Instead I solved both parts in Haskell, using the
knowledge that a substring may be repeated more than twice. The Haskell
solution is still a bit slow for my liking, but I'll port it to Uiua
before optimising.